refactor(post/photo) : add Entity and Status Enum
- create entity package in post
