+++ Photaiary/src/main/java/com/photaiary/Photaiary/post/photo/entity/DeleteStatus.java
@@ -1,0 +1,5 @@
+package com.photaiary.Photaiary.post.photo.entity;
+
+public enum DeleteStatus {
+    exist, not_exist
+}
